Fix dtype warning in stacked_dataset_builder person ID assignment
Cast np.arange output to int32 to match column dtype.

+++ policyengine_us_data/datasets/cps/local_area_calibration/stacked_dataset_builder.py
@@ -447,7 +447,9 @@ def create_sparse_cd_stacked_dataset(
             )
 
         # Create sequential IDs for this CD
-        new_person_ids = np.arange(start_id, start_id + n_persons_in_cd)
+        new_person_ids = np.arange(
+            start_id, start_id + n_persons_in_cd, dtype=np.int32
+        )
 
         # Assign all at once using loc
         combined_df.loc[cd_mask, person_id_col] = new_person_ids
